Added manager to menu and added some comments
