--- Editor/CustomEditors/AGXUnity+ObserverFrameRendererManagerEditor.cs
@@ -1,13 +1,0 @@
-
-using System;
-using AGXUnity;
-using AGXUnity.Collide;
-using UnityEditor;
-
-namespace AGXUnityEditor.Editors
-{
-  [CustomEditor( typeof( AGXUnity.ObserverFrameRendererManager ) )]
-  [CanEditMultipleObjects]
-  public class AGXUnityObserverFrameRendererManagerEditor : InspectorEditor
-  { }
-}
