+++ Editor/CustomEditors/AGXUnity+Rendering+ObserverFrameRendererManagerEditor.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 19ebaa55b82a9d84b948e814f298f8f2
+guid: f72a69f446077ff498e6c1dcfb4ccf19
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
